Refactor DEM calculation
